Add login with JWT methods

+++ invoices_gae/auth/authmanager.py
@@ -1,17 +1,24 @@
 #coding: utf-8
+import jwt
 import httplib2
 import json
 from flask import session
 
 
 class AuthManager:
-
+    JWT_SECRET = b'p\xbb+\xcfU\x06\xaf\x8e\xb6P60\x14\x88t(W\xfb:L\x82+\x0e\xa1\xc2\x11g`\xdf\xb3\xc0q'
+    JWT_PARAM_KEY_CREDENTIALS = 'payload'
+    SESSION_KEY_JWT = 'login_jwt'
     SESSION_KEY_PROFILE = 'profile'
 
     def __init__(self, oauth2):
-        oauth2.authorize_callback = self.login
+        """init
+
+        oauth2 is an instance of oauth2client.contrib.flask_util.UserOAuth2
+        """
+        if oauth2:
+            oauth2.authorize_callback = self.login_callback
         self.oauth2 = oauth2
-        self._instance = self
 
     def get_login_user(self):
         """Get current login user
@@ -20,18 +27,47 @@ class AuthManager:
         """
         return session.get(self.SESSION_KEY_PROFILE, None)
 
-    def login(self, credentials):
+    def get_login_user_by_jwt(self, jwt):
+        credentials = jwt.decode(encoded, key=JWT_SECRET, algorithms=['HS256'])[JWT_PARAM_KEY_CREDENTIALS]
+        return self._fetch_google_profile(credentials)
+
+    def login_callback(self, credentials):
+        """Callback function of OAuth2
+        """
+        profile = self._fetch_google_profile(credentials)
+        if profile is None:
+            return False
+        # generate jwt that containes credentials
+        encoded = jwt.encode({self.JWT_PARAM_KEY_CREDENTIALS: credentials.to_json()}, self.JWT_SECRET, algorithm='HS256')
+        # Save jwt in session to prevent other user fetch it
+        # Flask store data in cookie, not in server size. So store jwt in
+        # session will exceed max data limit of cookie.
+        # Must use Flas-Session to store session in service-side to make this
+        # function works.
+        session[self.SESSION_KEY_JWT] = encoded
+        session[self.SESSION_KEY_PROFILE] = profile
+        return True
+
+    def _fetch_google_profile(self, credentials):
+        """Fetch Google Plus Profile by Google API.
+
+        return dict that contains profile information, or return None if fetch failed.
+        """
         http = httplib2.Http()
         credentials.authorize(http)
         resp, content = http.request(
             'https://www.googleapis.com/plus/v1/people/me')
-
         if resp.status != 200:
-            return False
-        session[self.SESSION_KEY_PROFILE] = json.loads(content.decode('utf-8'))
-        return True
+            return None
+        return json.loads(content.decode('utf-8'))
+
+    def get_login_jwt(self):
+        """Get the previous generated jwt
+        """
+        return session.get(self.SESSION_KEY_JWT, None)
 
     def logout(self):
-        self.oauth2.storage.delete()
+        if self.oauth2:
+            self.oauth2.storage.delete()
         if self.SESSION_KEY_PROFILE in session:
             del session[self.SESSION_KEY_PROFILE]
